test: update odoo tests
