+++ config/settings.py
@@ -54,13 +54,10 @@ INSTALLED_APPS = [
     "references",
     "awg",
     "release",
-    "crm",
-    "crm.odoo",
+    "integrations",
     "post_office",
     "todos",
     "website",
-    "social.bsky",
-    "social.meta",
     "emails",
     "app",
     "django_celery_beat",
