+++ integrations/test_odoo.py
@@ -1,3 +1,15 @@
+import os
+import django
+
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings")
+django.setup()
+from django.conf import settings
+settings.ALLOWED_HOSTS=["testserver"]
+
+from django.core.management import call_command
+call_command("migrate", run_syncdb=True, verbosity=0)
+
+
 from django.test import Client, TestCase
 from django.urls import reverse
 from django.contrib.auth import get_user_model
@@ -21,7 +33,7 @@ class OdooTests(TestCase):
     def test_connection_success(self, mock_proxy):
         mock_srv = mock_proxy.return_value
         mock_srv.authenticate.return_value = 1
-        response = self.client.post(reverse("odoo-test", args=[self.instance.pk]))
+        response = self.client.post(reverse("integrations:odoo-test", args=[self.instance.pk]))
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.json()["detail"], "success")
 
@@ -29,8 +41,9 @@ class OdooTests(TestCase):
     def test_connection_invalid(self, mock_proxy):
         mock_srv = mock_proxy.return_value
         mock_srv.authenticate.return_value = False
-        response = self.client.post(reverse("odoo-test", args=[self.instance.pk]))
+        response = self.client.post(reverse("integrations:odoo-test", args=[self.instance.pk]))
         self.assertEqual(response.status_code, 401)
+
 
 class OdooAdminTests(TestCase):
     def setUp(self):
@@ -54,7 +67,7 @@ class OdooAdminTests(TestCase):
     def test_admin_action(self, mock_proxy):
         mock_srv = mock_proxy.return_value
         mock_srv.authenticate.return_value = 1
-        url = reverse("admin:odoo_instance_changelist")
+        url = reverse("admin:integrations_instance_changelist")
         resp = self.client.post(
             url,
             {
@@ -70,4 +83,3 @@ class OdooAdminTests(TestCase):
             self.instance.password,
             {},
         )
-
