+++ integrations/urls.py
@@ -2,10 +2,11 @@ from django.urls import path
 
 from . import views
 
-app_name = "bsky"
+app_name = "integrations"
 
 urlpatterns = [
     path("register/", views.register, name="register"),
     path("post/", views.post, name="post"),
     path("domain-post/", views.domain_post, name="domain-post"),
+    path("odoo/test/<int:pk>/", views.test_connection, name="odoo-test"),
 ]
