+++ integrations/views.py
@@ -29,3 +29,23 @@ def domain_post(request):
     text = request.POST["text"]
     post_from_domain(text)
     return JsonResponse({"status": "ok"})
+
+import xmlrpc.client
+from urllib.parse import urljoin
+from django.shortcuts import get_object_or_404
+
+from config.offline import requires_network
+from .models import Instance
+
+
+@requires_network
+def test_connection(request, pk):
+    instance = get_object_or_404(Instance, pk=pk)
+    server = xmlrpc.client.ServerProxy(urljoin(instance.url, "/xmlrpc/2/common"))
+    try:
+        uid = server.authenticate(instance.database, instance.username, instance.password, {})
+    except Exception as exc:  # pragma: no cover - network errors
+        return JsonResponse({"detail": str(exc)}, status=400)
+    if uid:
+        return JsonResponse({"detail": "success"})
+    return JsonResponse({"detail": "invalid credentials"}, status=401)
