+++ release/utils.py
@@ -82,8 +82,7 @@ def _write_pyproject(package: Package, version: str, requirements: list[str]) ->
                     "readme",
                     "website",
                     "release",
-                    "crm",
-                    "crm.odoo",
+                    "integrations",
                 ]
             }
         },
